feat: refactor logout functionality into separate LogoutButton component
